% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{drygate_palette}
\alias{drygate_palette}
\title{A Drygate brewery palette generator}
\usage{
drygate_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{SevenPeaks},\code{GladEyeIPA}, \code{KelvinPilsner}, \code{BearfaceLager}, \code{DiscoForkliftTruck}, \code{TakeMeTotheRiver},  \code{CrossingtheRubicon},
\code{CloudCity}, \code{Shred}, \code{Voller}, \code{Drygate},}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Drygate brewery.
}
\examples{
drygate_palette("SevenPeaks")
drygate_palette("GladEyeIPA")
drygate_palette("KelvinPilsner")
drygate_palette("BearfaceLager")
drygate_palette("DiscoForkliftTruck")
drygate_palette("TakeMeTotheRiver")
drygate_palette("CrossingtheRubicon")
drygate_palette("CloudCity")
drygate_palette("Shred")
drygate_palette("Voller")
drygate_palette("Drygate")


# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- drygate_palette(21, name = "DiscoForkliftTruck", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
