% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drygatebrewer.R
\docType{package}
\name{drygatebrewer-package}
\alias{drygatebrewer}
\alias{drygatebrewer-package}
\title{drygatebrewer: A Drygate brewery palette generator}
\description{
list of palettes from Drygate brewery
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MAClaassen/drygatebrewer}
  \item Report bugs at \url{https://github.com/MAClaassen/drygatebrewer/issues}
}

}
\author{
\strong{Maintainer}: M.A. Claassen \email{almuclaassen@gmail.com}

}
\keyword{internal}
