% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{drygate_palettes}
\alias{drygate_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
drygate_palettes
}
\description{
Use \code{\link{drygate_palette}} to construct palettes of desired length.
}
\keyword{datasets}
